Change file interaction functions
File access is consolidated into two functions, read_list() and write_list(),
for conversion between song lists and song list files.  Rank conversion is
consolidated to convert_list(), which converts a song list with ranks to a song
list.  main() is changed to use these new functions.  This is in preparation
for future "interactive" modes of operation, as well as general programming
goodness.

dl.py is also changed due to refactoring.
test/list2 is changed to reflect new output format.